requirements.txt
absl-py==0.11.0
astunparse==1.6.3
cached-property==1.5.2
cachetools==4.1.1
certifi==2020.11.8
chardet==3.0.4
cycler==0.10.0
Django==2.2.13
djangorestframework==3.9.2
gast==0.3.3
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.33.2
h5py==2.10.0
idna==2.10
importlib-metadata==2.0.0
joblib==0.17.0
jsonfield==3.1.0
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.3
matplotlib==3.3.3
more-itertools==8.6.0
music21==6.3.0
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.3.0
Pillow==8.0.1
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pygame==2.0.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.4
PyYAML==5.3.1
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.6
scipy==1.5.4
six==1.15.0
sqlparse==0.4.1
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-estimator==2.3.0
termcolor==1.1.0
urllib3==1.26.2
webcolors==1.11.1
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.0
